+++ apps/buildco-control/src/app/appearance.ts
@@ -1,11 +1,7 @@
 import minimal from "@capillaryjs/capillary-ui/themes/minimal/theme.css?url";
-import java from "@capillaryjs/capillary-ui/themes/java/theme.css?url";
+import capillary from "@capillaryjs/capillary-ui/themes/capillary/theme.css?url";
 import shiny from "@capillaryjs/capillary-ui/themes/shiny/theme.css?url";
-import original from "@capillaryjs/capillary-ui/themes/original/theme.css?url";
-import scifi from "@capillaryjs/capillary-ui/themes/scifi/theme.css?url";
 import soft from "@capillaryjs/capillary-ui/themes/soft/theme.css?url";
-import dark from "@capillaryjs/capillary-ui/themes/dark/theme.css?url";
-import glossy from "@capillaryjs/capillary-ui/themes/glossy/theme.css?url";
 import white from "@capillaryjs/capillary-ui/themes/white/theme.css?url";
 import gray from "@capillaryjs/capillary-ui/colors/gray/colors.css?url";
 import green from "@capillaryjs/capillary-ui/colors/green/colors.css?url";
@@ -17,7 +13,7 @@ import red from "@capillaryjs/capillary-ui/colors/red/colors.css?url";
 import yellow from "@capillaryjs/capillary-ui/colors/yellow/colors.css?url";
 
 export const themes = Object.entries({
-    minimal, java, shiny, original, scifi, soft, dark, glossy, white,
+    capillary, shiny, soft, white, minimal,
 }).map(([value, href]) => ({
     value,
     href,