+++ packages/capillary-ui/src/Components/lineinputs/textbox.tsx
@@ -133,11 +133,11 @@ export class Textbox extends LabeledInputControl<TextboxProps> {
             max-width: 100%;
             min-width: var(--input-min-width, 6rem);
             margin: 0;
-            padding: 0 5px;
+            padding: var(--ui-control-padding);
             color: var(--input-color);
             background: var(--input-background);
             border: var(--input-border);
-            border-radius: var(--radius-md);
+            border-radius: var(--ui-border-radius);
             box-shadow: var(--input-shadow);
             box-sizing: border-box;
             font: inherit;